+++ src/utils/helpers.py
@@ -4,10 +4,39 @@ import torch
 import torchaudio
 import numpy as np
 from pathlib import Path
+from argparse import ArgumentParser
 
 from data.dataset import SpringDataset
 from models.TCN import TCN
 from models.WaveNet import WaveNet
+
+
+def parse_args():
+    parser = ArgumentParser(description='Train a TCN model on the plate-spring dataset')
+    parser.add_argument('--datadir', type=str, default='../../datasets/plate-spring/spring/', help='Path (rel) to dataset ')
+    parser.add_argument('--audiodir', type=str, default='../audio/processed/', help='Path (rel) to audio files')
+    parser.add_argument('--logdir', type=str, default='../results/runs', help='name of the log directory')
+    parser.add_argument('--checkpoint_path', type=str, required=True, help='Path (rel) to checkpoint to load')
+
+    parser.add_argument('--device', type=str, 
+                        default="cuda:0" if torch.cuda.is_available() else "cpu", help='set device to run the model on')
+    parser.add_argument('--sample_rate', type=int, default=16000, help='sample rate of the audio')    
+    parser.add_argument('--n_epochs', type=int, default=2, help='the total number of epochs')
+    parser.add_argument('--batch_size', type=int, default=8, help='batch size')
+    parser.add_argument('--lr', type=float, default=0.001, help='learning rate')
+    parser.add_argument('--crop', type=int, default=3200, help='crop size')
+    
+    parser.add_argument('--max_length', type=float, default=None, help='maximum length of the output audio')
+    parser.add_argument('--stereo', action='store_true', help='flag to indicate if the audio is stereo or mono')
+    parser.add_argument('--tail', action='store_true', help='flag to indicate if tail padding is required')
+    parser.add_argument('--width', type=float, default=50, help='width parameter for the model')
+    parser.add_argument('--c0', type=float, default=0, help='c0 parameter for the model')
+    parser.add_argument('--c1', type=float, default=0, help='c1 parameter for the model')
+    parser.add_argument('--gain_dB', type=float, default=0, help='gain in dB for the model')
+    parser.add_argument('--mix', type=float, default=50, help='mix parameter for the model')
+    args = parser.parse_args()
+    
+    return parser.parse_args()
 
 
 def load_audio(input, sample_rate):
@@ -50,7 +79,7 @@ def load_data(datadir, batch_size):
 def initialize_model(device, model_type, hparams=None, criterion=None, checkpoint_path=None):
     """Initialize a specific model"""
     if device is None: 
-        device = torch.device("cuda:0" if torch.cuda.is_available() else "cpu")
+        device = torch.device("xpu")
     
         # Load the model from a checkpoint if provided
     if checkpoint_path is not None:
@@ -103,7 +132,7 @@ def initialize_model(device, model_type, hparams=None, criterion=None, checkpoin
 
 def save_model_checkpoint(model, hparams, criterion, optimizer, scheduler, n_epochs, batch_size, lr, timestamp):
     """Save the model if validation loss decreases"""
-    save_to = f'results/checkpoints/tcn_{n_epochs}_{batch_size}_{lr}_{timestamp}.pt'
+    save_to = f'../results/checkpoints/tcn_{n_epochs}_{batch_size}_{lr}_{timestamp}.pt'
     torch.save({
         'model_class': type(model),
         'model_state_dict': model.state_dict(),
@@ -116,4 +145,5 @@ def save_model_checkpoint(model, hparams, criterion, optimizer, scheduler, n_epo
     }, save_to)
 
     print(f"Checkpoint saved: {save_to}")
+
